OSS-fuzz #45030
The code to retrieve a page dictionary was assuming that, if we reached
a node in the pages tree which contained a dictionary that it was a
page dictionary. This file, however, has a node in the Pages tree which
has a Kids array, containing a reference back to a dictionary which
is the Root pf the Pages tree.

This commit checks the dictionary to see if it is already in the loop
detection array, if it is then we've already visited it, so this must
be a circular reference.

Also removed a couple of unused variables that were causing compiler
warnings.