Add pdfi_read_bare_keyword function.
Performs exactly the same as reading a token, but doesn't actually
allocate a pdf_obj for it, just returns an enum value. Saves on
allocating/deallocating keywords.

There are still places in the code that read keywords as objects,
hence pdf_keyword still exists. These will be dealt with in future
commits.