pdfi with Ghostscript - normalise 'Box' arrays before use.
After we retrieve a 'Box' from the PDF file we use it to setup the
device PageSize. However we were assuming the Box array was always in
the 'normal' format of bottom left corner and top right corner.

The spec is quite clear that any opposite pair is acceptable, so in this
commit we 'normalise' the two corners before we process the array for
the PageSize.

Also remove a commented out left over piece of debug printing.