Overprint simulation and OutputIntent interaction
In cases where -dOverprint=/simulate and -dUsePDFX3Profile
are used, if the page has overprint and no transparency push the
pdf14 device and render the page to a transparency buffer
to the Output intent color space.  If the target device has a
different ICC profile than the output intent (specified
with something like -sOutputICCProfile="Fogra39.icc") then
convert the output intent rendered buffer to the target
device color space.  Also make sure that the tiff device do
not use the wrong profile in the header.

Additionally, a problem was found where the pdf14 device
was setting up color conversions using the perceptual rendering
intent.  Those transforms should be using the colorimetric
rendering intent.