* Fix mismatch of page group push and pop condition
Thanks to Ken for finding this. If we are pushing the
pdf14 device we may be doing this if we are doing
overprint simulation or if there is transparency.
Only push the page group if we have Transparency
on the page (there are files which may have a
page group but never use transparency op_all.pdf is
an example of one).  Match the conditions for pushing
and popping the page group.  Here we only will push
the group if one exists, we have transparency on the
page and we need the pdf14 device (and the device push
and group push were successful).