Old PDF interpreter - improved CMap handling
The file supplied for this came from a customer who asked that we
delete the file after working on it, so there is no bug report and no
reproducer for this.

The problem is that the PDF file has a font using a CMap, and that CMap
uses /UseCMap (and the usecmap operator) to read a different (Horizontal)
CMap and then modify it with Vertical glyph positions.

The CMap does not have a begincodespacerange, it simply inherits the
ranges from the child CMap. This causes the code added to work around
Bug 690737 to read off the end of the CMap in read_CMap_stream.

Since there is nothing left to process, this causes errors in the CMap
processing.

Following a suggestion from Chris this commit first attempts the same
'discard up to the begincodespacerange) hackery, then checks to see if
the stream has any bytes left. If it does we proceed as before.

If there are no bytes left, then we have discarded all of the content.
So we rewind the stream to the point we were at before we tried to
discard the header, impose a different SubFileDecode, looking this time
for a 'begincmap' and then attempt to process the CMap as normal.