PDF interpreter - fix FitPage
When rewriting FitPage for the new interpreter I seem to have left part
of the debugging code in place, which led to two incorrect values on
the stack when we calculated the scale factor. I general these values
were zero, leading to a scale factor of 0 and nothing being drawn.

Fixed here by removing the extraneous values.