OSS-fuzz #45195 - check for memory allocation failure
pdf14_clist_push_color_model allocates a structure 'new_group_color' but
did not check to see if the allocation was successful before assigning
a value to one of the structure members.

If the allocation failed, this caused a crash.

Add a check to see if the allocation failed, and return an error if so.