Fix the VM attributes flags for the new PS device containers
When PS refs referenced devices directly, for the prototype devices (which are
not allocated on the heap) the object was marked as VM mode "avm_foreign", thus
the garbager and save/restore would not attempt to do anything it.

Now the refs reference a container, which is always allocated on the heap, that
attribute flag always needs to be the prevailing PS VM mode.

Discovered while investigating oss-fuzz: 45320