OSS_fuzz #45347 - validate image parameters
The file has an inline image with a /H (height) of -19. The interpreter
was simply passing that to the image rendering code which, not
unreasonably did nothing, but reported it had consumed -19 rows.

We then used that to try and skip over the image data, but treating
a negative number as an unsigned integer led to us trying to skip
ridiculously large amounts of data and eventually caused a crash in the
file handling.

This commit sets negative values to 0 (which then causes us to skip
the image) unless STOPONWARNING is set in which case we return an error.