oss-fuzz 45049: ASCII85Decode, avoid running off the end of the buffer
To maintain compatibility with Acrobat, the ASCII85Decode filter has to cope
with various breakages in the data, including whitespace between the '~' and
the '>', and even a missing '>'.

Searching the stream buffer for the '>', and not finding it, could leave the
pointer position a byte past the end of the buffer, this just ensures that
cannot happen.