Fix 'pget' for pdfi with Ghostscript
The utility program pdf2dsc.ps expects that both pget and knownoget will
be available in the current dictionary after runpdfbegin is executed.

To facilitate this pdfi makes such a definition but, unfortunately, it
had a bug if the key was not found in the dictionary, leaving a copy
of the dictionary on the operand stack when it should not.

Fixed here. Thanks to William Bader for spotting the bug.