Fix and extend chameleon device.
This wasn't working, possibly due to device changes, more likely
due to insufficient testing when it was first written.

We fix it here, and change the way we handle post render color
conversion.

We were using ICC before in the case when we had a PostRenderProfile,
but not otherwise. If no PostRenderProfile is specifically set, we
now stash the device profile before we change our colormodel, and
use that as the target for our post render conversion.

This has required a new downscaler helper function to make an
icc link from a given profile.

This means that we now only really support contone operation (as
doing ICC conversions from pre-dithered results will not look
good). If we want to support dithered output, then we'll really
need to do the dithering as part of the conversion from contone
rendered RGB, as otherwise it'll look really bad.