Rejig gsicc_set_device_profile.
Previously we'd allocate a profile structure, assign that, and
then go about filling it in. If we hit an error while filling it
in, we'd exit with the half filled structure still in force.

Rejig the code so that we allocate and fill the structure first,
and only assign it when it's well-formed.