Fix code to detect size of uncompressed stream data
When converting a stream to a buffer in memory, we need to know the size
of the uncompressed stream data, in order to allocate a buffer.

Chris improved the speed of this, but it turns out that the insane
stream code doesn't necessarily update the file pointer when we discard
a buffer, if the stream is just the native file.

So instead of using sreset(), use sbufskip(). Rather than simply
setting the buffer pointer back to the start, this sets it to the end,
as if we had actually consumed all the data. That then means
s_process_read_buf() will call stream_compact() and that will update
the stream 'position' member. Previously that wasn't happening, even
though we were reading more data from the file, which caused
considerable confusion!