OSS-fuzz #45508 - fix error path reference counting
The reference counts are correct, but after we've discarded the reference
held by pcs_alt (the alternate space for a /Separation space) we hadn't
set the pointer to NULL, which meant a subsequent error would decrement
pcs_alt again, leading to us freeing the colour space while the
ICC profile cache still had a reference to it.

Fix by simply setting pcs_alt to NULL in order to properly discard the
reference.