Bug 705063: Fix blank_unmasked_bits to work msb, not lsb.
When working with alphabits, we read pixels back from the destination
buffer, process it, and then write it out again. The reading of the
current pixels is done via get_bits_rectangle. When working with a
mask, this calls blank_unmasked_bits to avoid returning undefined
values.

This routine was incorrectly written to read the mask as it pixels
were packed into a byte least significant bit first. Correct it to
work most significant bit first.