Bug 705075: gpdl - Avoid unintended side effect of -dNODISPLAY
We need to initialise Ghostscript with the nullpage device when calling it via
gpdl (so we can later replace the nulldevice with the "real" output device).

Using the "shortcut" of "-dNODISPLAY" to install the nulldevice as the initial
device ends up disabling the interactive prompt between pages.

This just changes things specify "-sDEVICE=nullpage" explicitly, avoiding that
problem.