OSS-fuzz #45842 - avoid recursion to prevent C-stack overflow
pdfi_read_token() was calling itself recursively under a few conditions;
a closing procedure character '}', on encountering a comment, or if we
found a token delimiter unexpectedly and were not stopping on errors.

The OSS-fuzz file has a large number of string closing characters ')',
which is a delimiter and the recursion eventually overflowed. Recursion
is overly clever for this, we don't need recursion we can simply
consume the character(s) and loop back to the start of the function to
get the next token.