OSS-fuzz 45886 - validate funtion parameters for transfer functions
This was causing a buffer overrun when evaluating a transfer function
because the code assumed it would be a 1-in one-out function, whereas
the supplied function is a 1-in 3-out function, causing it to overrun
the output buffer.

This commit checks that the function is a 1-in, 1-out function and
returns an error if it isn't.

Similar fix applied for colour transfers, and a C++ commented out line
removed as it seemed to be entirely spurious.