pdfwrite - add controls for optional entries with variable content
Bug #704846 "Optionally omit CreatingTime in PDF output for reproducibility"

Some content in a PDF file is created using the system time, and will
vary depending on the precise time the file is created. This is entirely
correct and we are not prepared to alter this behaviour.

However said content is also often optional, but there was no way to
prevent pdfwrite creating it.

This commit adds three new controls; OmitDateInfo, OmitXMP and OmitID
which direct pdfwrite not to create and include the CreationDate and
ModDate in the /Info dict and Metadata (OmitInfoDate), the file ID in
the trailer dictionary (OmitID) and the XMP Metadata included in the
Catalog dictionary (OmitXMP).

There are, however, other considerations. XMP metadata is required for a
valid PDF/A file, /ID is required for PDF 2.0, and for encrypted files
and CreationDate is required for PDF/X-3.

As a result if you try to set these controls with one of those other
features, it will raise a warning and ignore the control.