pdfi - fix number parsing with malformed integers
Commit 4ef76e99d70f3f290120486434abeea24e7c6682 altered the way number
parsing takes place for performance purposes. While that worked well
for properly formed numbers it introduced a problem with malformed
integers.

The previous code, on encountering a malformed number (eg 50+20)
would replace the '+' in the buffer with a NULL and continue reading
as long as there were digits to read, storing them behind the NULL.

It would then use the buffer to find the required value. Because the
digits after the error were stored after the NULL they didn't get used
for the value.

The new code accumulates integer values as we go, so by continuing to
read digits after the error, they were considered as part of the value
which resulted in an incorrect parsed value.

In addition, if we encountered a '-' in the wrong place (eg 50-20) then
we would similarly ignore it, but the current scanning takes it into
account, again leading to an incorrect value.

This commit simply skips these if 'malformed' is true and 'recovered'
is also true. Together these indicate we are skipping to the end of the
numeric data and ignoring it.