High-level patterns - correct use of ID
When handling high level patterns we were either casting a gs_id to an
int, or passing a gs_id as an int. Since a gs_id is an unsigned long
this potentially causes problems on any platform where the two types
differ in size (eg Linux).

Fixed here by using the id directly from the pattern instance for the
gxdso_pattern_start_accum case, and by passing the id as the data
parameter rather than the size parameter for the gxdso_pattern_load
case.