-dBlackVector=true/false option added
If set to true, stroke and fill operations that are fills
of DeviceGray, DeviceRGB, DeviceCMYK, ICC, or LAB colors
will be black.  The exception is that white fills will remain white.

DeviceN, Separation, shading and pattern fills will remain as is.

We use the same structure as the -dBlackText option used to store
the color space into the graphic state.  This structure has been renamed
black_textvec_state from black_text_state to denote this usage, as have
other variables.