pdfwrite - fix group attributes ColorSpace with old PDF interpreter
Commit 4659f3aedb77a039e54e4e1f3038cafa0c10a283 altered pdfwrite to use
the ColorSpace passed as a parameter, instead of relying on the
interpreter to set the current colour space. However I missed one case
with the old PostScript-based PDF interpreter where it specifically
sets the ColorSpace in the params to be NULL (begingtranspaencymaskgroup).

Because this code specifically sets the member to be NULL I presume this
has some significance in the transparency code (no idea what though) so
I don't want to change it. Instead alter pdfwrite so that if the
structure member is NULL we fall back to using the current colour space.