pdfwrite - use consistent data types when creating PDF/X
We were using double for the box variables and float for the temporary
variables used to convert the strings. This doesn't cause problems
unless we end up copying from one data type to another where we can
lose precision causing errors.

Using floats seems good enough and this is a simpler fix than replacing
all the uses of floats with doubles.