GhostPDF - fix use of .PDFFile with other operators
I believe this originally worked, but later changes have caused it to
stop working. When we use .PDFFile to feed a filename to the PDF
interpreter for processing we don't have any way to detect it. The other
operators (eg .PDFInfo) check to see if there is an open file associated
with the context, and give an ioerror if there isn't. I suspect this was
added later, and because there is no clear way to tell that .PDFFile has
been used (we can detect use of .PDFStream) the operation failed.

This commit just adds a flag which is set if .PDFFile successfully
opens a PDF file by name, and then checks that flag in addition to
the pdf_stream to ensure it is safe for the other operators to execute.