Implement error recovery in gsicc_initialize_iccsmask
OSS-fuzz bug #46425

The problem here is that the file is truly knackered. What happens is
that we have an image SMask which calls gs_begin_transparency_mask().

That determines that the 'smask_profiles'  in the ICC manager attached
to the graphics state have not yet been initialised (they are NULL) and
calls gsicc_initialize_iccsmask() to do so.

gsicc_initialize_iccsmask() creates the smask_profiles structure, and
then calls gsicc_set_iccsmaskprofile() to set up each of the Gray, RGB
and CMYK profiles in turn. The Gray profile fails, and we return an
error. Crucially, however, we do not free 'smask_profiles'.

Then we encounter another image SMask, that again calls
gs_begin_transparency_mask(). This time the graphics state ICC manager
has an entry for smask_profiles (the aborted previous one) so we don't
try and set it up, we just try and use those profiles, Since they are
NULL this leads to a seg fault.

Fixed here by freeing the 'smask_profiles' structure and setting the
member of the ICC manager back to NULL, if the creation of a profile
fails (we also free any profiles that did succeed).