Fix SEGV in -ZL handling caused by debug code.
When keeping track of where command blocks start in the
clist offset map (only used for debugging), we can get
confused if we happen to have absorbed exactly all the
bytes in the current buffer (i.e. buffer_offset == 4096).

We used to return -1 in this case which would cause
SEGVs.