GhostPDF - Don't abort the page with errors on /Group
Bug #705026 "Error when reading latex-pdf"

The PDF file has a page dictionary which has a /Group key where the
value is an indirect reference to a free object. Previously we would
abort the page processing at that point, this commit instead flags the
error but continues to process the page Content stream.