OSS-fuzz #465222 - ensure devices are closed to prevent memory leak
Bug #705211 "OSS-fuzz bug #46522"

The problem is (superficially) caused by pdfi returning an invalid
number in the media request for the second page.

Because we have already imaged the first page, we have called code in
cups_print_pages() which allocates memory via cupsRasterOpen(). That
memory is released when we close the device. This seems reasonable; we
can't have reached this point without opening the device and it should
be closed before we exit.

However, because of the illegal media, clist_open() fails, because
clist_init() fails. The function clist_open() sets the device 'is_open'
member to false before it begins, but in the event of an error it does
not restore it, this means that when the device goes out of scope it is
no longer open, so we don't call the device's close method so we leak
the memory.

A similar problem appears to exist in clist_mutate_to_clist(), which is
where clist_open() is called from.

This commit changes both cases so that if the device was open, it
remains open. That then causes seg faults.

The seg fault is a result of gdev_prn_allocate(); if an error occurs
(ie from the clist), the it frees the 'print device' bg_print pointer,
but it doesn't set the pointer to NULL. Later we try to tear down the
'print device' and because bg_print is not NULL we try to use it in
prn_finish_bg_print() and that causes a seg fault.

Fix that by setting bg_print to NULL when we free it. In two places....