pdfwrite - try to resolve double free error
Move the code to free the named objects before we free the 'Other'
resources. At least some of the named resources can be stored as
resources as well, and if we free them as resources before we free
them from the named resources objects we end up trying to access a
freed object.