Bug 705221: Fix explosion in memory use with given file.
The new PDF interpreter uses the chunk memory manager, rather than
the clump based one used by the Postscript based one. Accordingly,
it was falling into a bug with this memory manager.

Any block larger than a threshold (32768 bytes) is allocated in its
own block. On allocation, we figure that block size as including
the header we put on the block. On deallocation, we were failing to
include that.

Accordingly blocks were being seen as large on allocation, and small
on deallocation, resulting in them 'falling through the cracks'.

This commit fixes that, and tweaks some of the debugging code. The
file behaves nicely with this in.