Extend BlackVector to cope with rectfill
The PostScript rectfill operator does not go through the usual path for
a general fill, instead it takes an optimised route (because the path
is guaranteed to be rectangular and therefore scan-converts trivially).

The BlackVector code was not implemented on that path, which meant it
didn't have any effect on code using rectfill.

This commit simply duplicates the approach taken in do_fill() into the
rectfill code path.