Bug 705221(redux): "Reverse" original fix
Original fix was to make the chunk freeing code to identify single object
chunks match the allocating code - i.e. both using the "padded" size.

This appears to cause confusion with other areas in the code, leading to
potential segfaults.

This reverses that change, so both the allocating and freeing code use the
unpadded size.

This still solves the memory explosion bug, but doesn't cause the issues
identified by oss-fuzz 46830 and 46734.

Ultimately it may be preferable to revise the rest of the code to avoid the
confusion, but this avoids a flood of issues from oss-fuzz.