oss-fuzz 46878: Don't free enumerator on tint transform error
In the event of a catastrophic error while executing the function for a tint
transform, we'd free the enumerator before returning the error from the
"_continue" function.

Turns out, the enumerator may still be referenced on the exec stack, meaning
we suffer a heap-use-after-free error when the garbage collector later does
its mark pass on the Postscript stacks.

Don't free the enumerator on this type of error, and leave it to the garbager
to deal with it.