Bug 705175    Index sep color space with /None entry
Thanks to Ken for finding much of this. When we had an indexed color
space whose base was a Separation of DeviceN colorspace with
all /None entries, we were not setting the device color to null
which would indicate that no color is drawn.  This problem occurred
dealing with images and with graphics.