Bug 705250: Maintain glyph name -> GID -> charstring mapping
In Type 2 fonts, we were "collapsing" the mapping of the glyph name to GID to
charstring mapping into a single glyph name to charstring map. That is contrary
to the description in the PLRM 3rd Edition.

This becomes a problem when attempting to re-encode the font by replacing the
glyph name to GID encoding in the CharStrings dictionary, where we expect to
find a charstring, we find an integer (GID).

This changes how we create Type 2 fonts from CFF so it maintains that two
step mapping, and re-encoding will now work.