Bug 705253: Add equivalent glyph name mappings
The old interpreter, when loading a Type 1 font for substitution would,
for every key in the CharStrings dictionary, search the Adobe Glyph list,
if a matching glyph name was found, it would note the Unicode code point,
and then search the AGL for any other glyph names with the same Unicode code
point.

We had hoped to avoid that with the new PDF interpreter, but it turns out we
cannot.

In the new code, we don't add those alternatives to the CharStrings dictionary
on creation, but do the mapping lazily on the event the input attempts to use
a glyph name that doesn't already exist.