GhostPDF - handle comments in type 4 calculator functions
No bug report or reproducer for this, the original files are hundreds
of megabytes.

The specification doesn't mention comments in type 4 calculator
functions, and it makes it clear that these are neither real PostScript
nor PDF streams, so we didn't permit comments.

However Acrobat obviously does permit these, so add code here to simply
ignore them.