Bug 705263 : Improper ICC profile used with highlevel clist image
When a high level image is written into the clist, values that are in the
image settings that are different that the clist device settings
are copied and a flag is set to indicate that those values need to be
written to the clist with the writing of the unknown command (cmd_write_unknown).
A problem is that the test for the change in the color space in the image
code did not test if the ICC profile
was different than the one stored in the clist device.  And if the color
spaces were different, the ICC information was not getting updated.