GhostPDF - ignore errors in setting halftones
Originally reported against the old PDF interpreter in bug #702077, the
file has a ridiculous halftone which has a Frequency of 1, at higher
resolutions we cannot create a halftone cell with such a low frequency
and generate an error.

Originally we altered the old PostScript-based to note the error and
carry on. GhostPDF does the same, but aborts the entire GState sequence
leading to missing content.

This commit instead swallows the graphics library error and substitutes
a PDF-specific error, so that the error is not completely ignored.
If PDFSTOPONERROR is set then we do return the underlying error.